Version 2 of our parkvibe
New add-in features:

1. delete photo function:
Admin can delete every photo in every OI;
User can only delete the photos they uploaded.
problem:
the delete buttons are all visible to admin/user, we cannot hide the delete buttons of the photos not belong to the users,  because of the restrictions of stagger recycle view

2. Create OIs according to the sg dataset
The application will create 20 OIs with the info is read from the dataset and listed in the current main page when click mapDisplay.
problem:
the description info is not included because the description read from dataset is not clean.